Simplify integration test infra
