--- .integration/src/storage.rs
@@ -1,113 +1,0 @@
-// Copyright (C) 2019-2021 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-use crate::Ledger;
-use snarkvm_dpc::{
-    block::Block,
-    traits::{LedgerScheme, Parameters, Storage, TransactionScheme},
-    TransactionError,
-};
-use snarkvm_utilities::{FromBytes, ToBytes};
-
-use rand::{thread_rng, Rng};
-use std::io::{Read, Result as IoResult, Write};
-
-pub fn random_storage_path() -> String {
-    let random_path: usize = thread_rng().gen();
-    format!("./test_db-{}", random_path)
-}
-
-// Initialize a test blockchain given genesis attributes
-pub fn initialize_test_blockchain<C: Parameters, T: TransactionScheme, S: Storage>(
-    genesis_block: Block<T>,
-) -> Ledger<C, T, S> {
-    let mut path = std::env::temp_dir();
-    path.push(random_storage_path());
-
-    Ledger::new(Some(&path), genesis_block).unwrap()
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TestTx;
-
-impl TransactionScheme for TestTx {
-    type Commitment = [u8; 32];
-    type Digest = [u8; 32];
-    type EncryptedRecord = [u8; 32];
-    type InnerCircuitID = [u8; 32];
-    type Memorandum = [u8; 64];
-    type SerialNumber = [u8; 32];
-    type Signature = [u8; 32];
-    type ValueBalance = i64;
-
-    fn transaction_id(&self) -> Result<[u8; 32], TransactionError> {
-        Ok([0u8; 32])
-    }
-
-    fn network_id(&self) -> u8 {
-        0
-    }
-
-    fn ledger_digest(&self) -> &Self::Digest {
-        &[0u8; 32]
-    }
-
-    fn inner_circuit_id(&self) -> &Self::InnerCircuitID {
-        &[0u8; 32]
-    }
-
-    fn old_serial_numbers(&self) -> &[Self::SerialNumber] {
-        &[[0u8; 32]; 2]
-    }
-
-    fn new_commitments(&self) -> &[Self::Commitment] {
-        &[[0u8; 32]; 2]
-    }
-
-    fn value_balance(&self) -> i64 {
-        0
-    }
-
-    fn memorandum(&self) -> &Self::Memorandum {
-        &[0u8; 64]
-    }
-
-    fn signatures(&self) -> &[Self::Signature] {
-        &[[0u8; 32]; 2]
-    }
-
-    fn encrypted_records(&self) -> &[Self::EncryptedRecord] {
-        &[[0u8; 32]; 2]
-    }
-
-    fn size(&self) -> usize {
-        0
-    }
-}
-
-impl ToBytes for TestTx {
-    #[inline]
-    fn write_le<W: Write>(&self, mut _writer: W) -> IoResult<()> {
-        Ok(())
-    }
-}
-
-impl FromBytes for TestTx {
-    #[inline]
-    fn read_le<R: Read>(mut _reader: R) -> IoResult<Self> {
-        Ok(Self)
-    }
-}
